Merge in NWS/ea_nws_website from feature/EAN-1205-organisation-multiple-admin-logins to develop-option-2

* commit '90bc2cf164b5e3d19db0ecd1c8c513aadd0a5e3b':
  Hooked up invite page to sign in admin EAN-1205
  hooked up contact promotion EAN-1205
  Added backend routes for changing admins EAN-1205
  Updated to 0.3.9-patch6
  Fixed duplicate check for shapefiles
  Added active admins to elasticache EAN-1205@

+++ backend/server/services/elasticache.js
@@ -552,8 +552,25 @@ const removeLinkedContacts = async (client, orgId, locationID, contactIDs) => {
   }
 }
 
-const orgSignIn = async (client, profile, organization, locations, contacts) => {
+const addOrgActiveAdmins = async (client, orgId, authToken) => {
+  const key = orgId + ':org_active_admins'
+  await addToList(client, key, authToken)
+}
+
+const removeOrgActiveAdmins = async (client, orgId, authToken) => {
+  const key = orgId + ':org_active_admins'
+  await removeFromList(client, key, authToken)
+}
+
+const getOrgActiveAdmins = async (client, orgId) => {
+  const key = orgId + ':org_active_admins'
+  const activeAdmins = await getList(client, key)
+  return activeAdmins
+}
+
+const orgSignIn = async (client, profile, organization, locations, contacts, authToken) => {
   await setJsonData(client, profile.id + ':profile', profile)
+  await addOrgActiveAdmins(client, organization.id, authToken)
   const orgExists = await checkKeyExists(client, organization.id + ':org_data')
   if (orgExists) {
     const existingLocations = await getLocationKeys(client, organization.id)
@@ -585,42 +602,48 @@ const orgSignIn = async (client, profile, organization, locations, contacts) => 
   }
 }
 
-const orgSignOut = async (client, profileId, orgId) => {
-  // delete all data from elasticache
+const orgSignOut = async (client, profileId, orgId, authToken) => {
   // delete profile
   await deleteJsonData(client, profileId + ':profile')
-  // delete org
-  await deleteJsonData(client, orgId + ':org_data')
-  // delete locations
-  // delete contacts
-  const locationKeys = await getLocationKeys(client, orgId)
-  const invLocationKeys = await getInvLocationKeys(client, orgId)
-  const contactKeys = await getContactKeys(client, orgId)
-  await Promise.all(
-    locationKeys.map(async (key) => {
-      await deleteJsonData(client, key)
-    })
-  )
-  await Promise.all(
-    invLocationKeys.map(async (key) => {
-      await deleteJsonData(client, key)
-    })
-  )
-  await Promise.all(
-    contactKeys.map(async (key) => {
-      await deleteJsonData(client, key)
-    })
-  )
+  await removeOrgActiveAdmins(client, orgId, authToken)
+  // get the new list to see if there are any admins still logged in
+  const activeAdmins = await getOrgActiveAdmins(client, orgId)
+  // delete all data from elasticache if there are no org Admins logged in
+  if (activeAdmins.length === 0) {
+    // delete org
+    await deleteJsonData(client, orgId + ':org_data')
+    // delete locations
+    // delete contacts
+    const locationKeys = await getLocationKeys(client, orgId)
+    const invLocationKeys = await getInvLocationKeys(client, orgId)
+    const contactKeys = await getContactKeys(client, orgId)
+    await Promise.all(
+      locationKeys.map(async (key) => {
+        await deleteJsonData(client, key)
+      })
+    )
+    await Promise.all(
+      invLocationKeys.map(async (key) => {
+        await deleteJsonData(client, key)
+      })
+    )
+    await Promise.all(
+      contactKeys.map(async (key) => {
+        await deleteJsonData(client, key)
+      })
+    )
 
-  await deleteData(client, orgId + ':t_POIS_locID')
-  await deleteData(client, orgId + ':t_invPOIS_locID')
-  await deleteData(client, orgId + ':t_Contacts_ID')
-  // delete contact and location keywords
-  await deleteJsonData(client, orgId + ':t_Keywords_location')
-  await deleteJsonData(client, orgId + ':t_Keywords_contact')
-  await deleteJsonData(client, orgId + ':alertLocations')
-  await deleteJsonData(client, orgId + ':t_Linked_locations')
-  await deleteJsonData(client, orgId + ':t_Linked_contacts')
+    await deleteData(client, orgId + ':t_POIS_locID')
+    await deleteData(client, orgId + ':t_invPOIS_locID')
+    await deleteData(client, orgId + ':t_Contacts_ID')
+    // delete contact and location keywords
+    await deleteJsonData(client, orgId + ':t_Keywords_location')
+    await deleteJsonData(client, orgId + ':t_Keywords_contact')
+    await deleteJsonData(client, orgId + ':alertLocations')
+    await deleteJsonData(client, orgId + ':t_Linked_locations')
+    await deleteJsonData(client, orgId + ':t_Linked_contacts')
+  }
+  
 }
 
 module.exports = {
@@ -649,5 +672,7 @@ module.exports = {
   addLinkedContacts,
   removeLinkedContacts,
   orgSignIn,
-  orgSignOut
+  orgSignOut,
+  checkKeyExists,
+  addOrgActiveAdmins
 }
